Pause Element selecton while CTRL pressed
This is a VERY NEEDED FEATURE and it solves a very very annoying thing
that was happening in element selection which was that clicking
on certain element triggered some action in the website like
menu bar opening, etc.
which makes it impossible for the extension to select that item.
Now with this feature one can simply hover over those kind of elements
and press Control key, which will pause the element selection
and allow the person to click on the newly added Select Element button
in the element selection div to Select it!!
I am very happy because of this as it has given me hope that this
garbage extension CAN ACTUALLY WORK IN REAL WORLD!!
I jumped with excitement when I added this feature and solve that
problem...
Love you Kanha!