Major Improvements and changes for setter.js
- The CSS code is improved to have
default styles and variables in order to avoid conflicts with
the webpage styles
- Changes to some class naming conventions
- The innerHTML for the elements that were appended to DOM are no longer
put in a variable in setter.js rather they are
fetched from HTML files. This makes
editing HTML a lot easier and less syntax errors will occur.
- The mws-element class will no longer be put in every element INSIDE
the HTML, rather, there is a function used to loop through
all the children and add the class to them.
- The UI for element selector is complete changed and improved
for better UI and UX.