fix: Font awesome not visible + element styling issue
In elementSelectors.css, the styles for details and
summary elements were declared without any safe measure and
therefore when the styles were injected into any webpage, the styles
of that page's details and summary elements were affected.

Second issue was that font-awesome icons weren't visible, so I have
included the required files in the manifest to solve it.