Version 1.0.0 first draft
Yes A WHOLE LOT OF CHANGES,
But I am really tired and frustrated now.
There are so many changes that I can't make individual commits now.
Here are some of the major changes:

**GLOBAL**
- There are now titles for almost all the important UI elements.
- Many new quick shortcuts are added
- New Sound Effects are added and more elements now
have sound effects
- Many Functionality bugs & style issues are solved
- Now the existing shortcuts are extracted not just of the website,
but also of the domain, and page (if full path) that are available for
the current website. Custom urls
are also tested but the functionality is not complete so custom urlType
is disabled rn.

**OPTIONS PAGE**
The whole UI is revamped,
- There are sub urls now under each domain and
not separate URLs. So pages are now listed inside a domain
list item.
- The About page is built very well now, it was
blank previously
- There is a help button in the bottom left  with
a Helper dialog having the support links
-  Footer is changed to created by and copyright text
-  Shortcut list is having a little box to show the urlType
and action of the shortcut which appears on hover
just like edit & delete buttons,
- Shortcut Action can be edited now
- Shortcut edit urlType functionality is not built completely,
so it's element is hidden as of now. It will come in next version.
-  Lights don't do complete darkness when both are off
- Disable for website is changed to disable for domain, domain disabled
means no shortcuts of any page
inside it will work.

**POPUP**
- Open options button styles are changed a lot.
-  Current Shortcuts shows a total shortcuts counter
- Disable for site is changed to disable
for domain, if domain is disabled,
all the pages inside it will be disabled too.
- A new domain indicator in the Disable Domain toggle text

**ELEMENT SELECTOR**
- Confirm Selection button will only appear when element selection is
paused.
- A Little details summary element in the bottom of element selector
which tells 3 ways to select an element.

**KEY SELECTOR**
- A whole of new Url Types are added and are working now.
- New Focus Action is added
- Custom URL editor is built but the functionality is having some issues
so it's not available in this version.

**useCustomShortcut**
- New Focus Action functionality added.
- Force shortcut action on double key press. If a shortcut key is
pressed two times consecutively & quickly, the action will be performed
without going through any conditionals. So even if an Input is active,
the action will be done.