fix: Allow full path if no page but only domain
Due to some flaw in conditions, key selector wasn't
enabling full path url type if the
url was having no pages
and just the domain.
But on some sites like Kindle books,
there is a domain & a search parameter that comes when opening a book.
So for those kinda sites too this option is now visible