fix: crashing and slowing system on URL changes
When the URL was changing, there were multiple setIntervals getting set
as it was inside the setCurrentURL function which was actually called inside
that setInterval which was getting called every 500 ms and called
setCurrentURL and getting data again from storage
every time url Was changing.

First of all, getting complete data from storage didn't made sense as it
is supposed to be fetched just once in a tab.
So I have separated the function to get Complete
data from storage and removed the get complete data
function from the updateData function.

And now the setInterval is set in the init so that there isn't multiple
of them.